2015 - Add template classes for all solutions and tests
Currently, when I go to solve a new puzzle, I have to create the packages and classes for solution and test cases each time before I start. By creating the packages and template classes ahead of time it will reduce the effort taken to get started for future puzzle solves. The main downside to this approach is that it becomes impossible to tell the difference between days that have been solved or not when looking at the file structure.

In the future, adding methods to automatically update README files with the "stars achieved" for each puzzle could be an approach to tracking the puzzles that have been solved.