Improve the method of determining which solution to run.
Instead of listing all the runnable solutions, which is tedious and error-prone, the method now determines which solution to run based on the solution's classpath.

The downside to this approach is that the current classpath for solutions does not support this approach without extra overhead, and the packages for each day will need to be renamed.